Fix use of gradient hooks (#23058)
* Fix way prop insertBlockAfter is send.

* Use HOC to wrap around gradient hook.

* Improve gradient HOC to be more generic.

* Revert "Fix way prop insertBlockAfter is send."

This reverts commit e0d921dcf059aa0169dbf01efb10d7c797f6fbd2.

* Make withGradient available to RN libs.

* Move export of withGradient to the gradient folder.
