+++ packages/block-editor/src/components/gradients/index.native.js
@@ -1,0 +1,2 @@
+export * from './use-gradient';
+export * from './with-gradient';
